test: consolidate Python client tests (#49450)
issue: #49202

This PR consolidates Python client tests by moving the remaining
milvus_client_v2 search tests into the main milvus_client suite and
removing the milvus_client_v2 folder.

It also reduces repeated collection/index setup in the data integrity
expression test by looping through expression fields within a single
collection setup.

Verification:
- python3 -m py_compile on touched Python test files
- PYTHONPATH=tests/python_client python3 -m pytest -c /dev/null
--collect-only -q for migrated tests and the refactored data integrity
test

---------
