+++ tests/python_client/milvus_client/test_milvus_client_data_integrity.py
@@ -75,19 +75,14 @@ class TestMilvusClientDataIntegrity(TestMilvusClientV2Base):
     @pytest.mark.parametrize("is_flush", [True])
     @pytest.mark.parametrize("is_release", [True])
     @pytest.mark.parametrize("single_data_num", [50])
-    @pytest.mark.parametrize("expr_field", [ct.default_int64_field_name,
-                                            ct.default_string_field_name,
-                                            ct.default_float_array_field_name])
     def test_milvus_client_query_all_field_type_all_data_distribution_all_expressions_array(self,
                                                                                             enable_dynamic_field,
                                                                                             supported_numeric_scalar_index,
-                                                                                            supported_varchar_scalar_index,
                                                                                             supported_json_path_index,
                                                                                             supported_array_double_float_scalar_index,
                                                                                             is_flush,
                                                                                             is_release,
-                                                                                            single_data_num,
-                                                                                            expr_field):
+                                                                                            single_data_num):
         """
         target: test query using expression fields with all supported field type after all supported scalar index
                 with all supported basic expressions
@@ -175,26 +170,32 @@ class TestMilvusClientDataIntegrity(TestMilvusClientV2Base):
         if is_flush:
             self.flush(client, collection_name)
         # 4. query when there is no index under all expressions
-        express_list = cf.gen_field_expressions_all_single_operator_each_field(expr_field)
-        compare_dict = {}
-        for i in range(len(express_list)):
-            json_list = []
-            id_list = []
-            log.info(f"query with filter '{express_list[i]}' before scalar index")
-            res = self.query(client, collection_name=collection_name,
-                             filter=express_list[i], output_fields=["count(*)"])[0]
-            count = res[0]['count(*)']
-            # log.info(f"The count(*) after query with filter '{express_list[i]}' before scalar index is: {count}")
-            res = self.query(client, collection_name=collection_name,
-                             filter=express_list[i], output_fields=[f"{expr_field}"])[0]
-            for single in res:
-                id_list.append(single[f"{default_primary_key_field_name}"])
-                json_list.append(single[f"{expr_field}"])
-            assert count == len(id_list)
-            assert count == len(json_list)
-            compare_dict.setdefault(f'{i}', {})
-            compare_dict[f'{i}']["id_list"] = id_list
-            compare_dict[f'{i}']["json_list"] = json_list
+        expr_fields = [ct.default_int64_field_name,
+                       ct.default_string_field_name,
+                       ct.default_float_array_field_name]
+        compare_dict_by_field = {}
+        for expr_field in expr_fields:
+            express_list = cf.gen_field_expressions_all_single_operator_each_field(expr_field)
+            compare_dict = {}
+            for i in range(len(express_list)):
+                json_list = []
+                id_list = []
+                log.info(f"query field '{expr_field}' with filter '{express_list[i]}' before scalar index")
+                res = self.query(client, collection_name=collection_name,
+                                 filter=express_list[i], output_fields=["count(*)"])[0]
+                count = res[0]['count(*)']
+                # log.info(f"The count(*) after query with filter '{express_list[i]}' before scalar index is: {count}")
+                res = self.query(client, collection_name=collection_name,
+                                 filter=express_list[i], output_fields=[f"{expr_field}"])[0]
+                for single in res:
+                    id_list.append(single[f"{default_primary_key_field_name}"])
+                    json_list.append(single[f"{expr_field}"])
+                assert count == len(id_list)
+                assert count == len(json_list)
+                compare_dict.setdefault(f'{i}', {})
+                compare_dict[f'{i}']["id_list"] = id_list
+                compare_dict[f'{i}']["json_list"] = json_list
+            compare_dict_by_field[expr_field] = compare_dict
         # 5. release if specified
         if is_release:
             self.release_collection(client, collection_name)
@@ -250,35 +251,38 @@ class TestMilvusClientDataIntegrity(TestMilvusClientV2Base):
             # 10. sleep for 60s to make sure the new index load successfully without release and reload operations
             time.sleep(60)
         # 11. query after there is index under all expressions which should get the same result
-        for i in range(len(express_list)):
-            json_list = []
-            id_list = []
-            log.info(f"query with filter '{express_list[i]}' after index")
-            count = self.query(client, collection_name=collection_name, filter=express_list[i],
-                               output_fields=["count(*)"])[0]
-            # log.info(f"The count(*) after query with filter '{express_list[i]}' after index is: {count}")
-            res = self.query(client, collection_name=collection_name, filter=express_list[i],
-                             output_fields=[f"{expr_field}"])[0]
-            for single in res:
-                id_list.append(single[f"{default_primary_key_field_name}"])
-                json_list.append(single[f"{expr_field}"])
-            # if len(json_list) != len(compare_dict[f'{i}']["json_list"]):
-            #     log.debug(
-            #         f"the field {expr_field} value after indexed under expression '{express_list[i]}' is:")
-            #     log.debug(json_list)
-            #     log.debug(
-            #         f"the field {expr_field} value before index to be compared under expression '{express_list[i]}' is:")
-            #     log.debug(compare_dict[f'{i}']["json_list"])
-            assert json_list == compare_dict[f'{i}']["json_list"]
-            # if len(id_list) != len(compare_dict[f'{i}']["id_list"]):
-            #     log.debug(
-            #         f"primary key field {default_primary_key_field_name} after indexed under expression '{express_list[i]}' is:")
-            #     log.debug(id_list)
-            #     log.debug(
-            #         f"primary key field {default_primary_key_field_name} before index to be compared under expression '{express_list[i]}' is:")
-            #     log.debug(compare_dict[f'{i}']["id_list"])
-            assert id_list == compare_dict[f'{i}']["id_list"]
-            log.info(f"PASS with expression {express_list[i]}")
+        for expr_field in expr_fields:
+            express_list = cf.gen_field_expressions_all_single_operator_each_field(expr_field)
+            compare_dict = compare_dict_by_field[expr_field]
+            for i in range(len(express_list)):
+                json_list = []
+                id_list = []
+                log.info(f"query field '{expr_field}' with filter '{express_list[i]}' after index")
+                count = self.query(client, collection_name=collection_name, filter=express_list[i],
+                                   output_fields=["count(*)"])[0]
+                # log.info(f"The count(*) after query with filter '{express_list[i]}' after index is: {count}")
+                res = self.query(client, collection_name=collection_name, filter=express_list[i],
+                                 output_fields=[f"{expr_field}"])[0]
+                for single in res:
+                    id_list.append(single[f"{default_primary_key_field_name}"])
+                    json_list.append(single[f"{expr_field}"])
+                # if len(json_list) != len(compare_dict[f'{i}']["json_list"]):
+                #     log.debug(
+                #         f"the field {expr_field} value after indexed under expression '{express_list[i]}' is:")
+                #     log.debug(json_list)
+                #     log.debug(
+                #         f"the field {expr_field} value before index to be compared under expression '{express_list[i]}' is:")
+                #     log.debug(compare_dict[f'{i}']["json_list"])
+                assert json_list == compare_dict[f'{i}']["json_list"]
+                # if len(id_list) != len(compare_dict[f'{i}']["id_list"]):
+                #     log.debug(
+                #         f"primary key field {default_primary_key_field_name} after indexed under expression '{express_list[i]}' is:")
+                #     log.debug(id_list)
+                #     log.debug(
+                #         f"primary key field {default_primary_key_field_name} before index to be compared under expression '{express_list[i]}' is:")
+                #     log.debug(compare_dict[f'{i}']["id_list"])
+                assert id_list == compare_dict[f'{i}']["id_list"]
+                log.info(f"PASS with field {expr_field} and expression {express_list[i]}")
         self.drop_collection(client, collection_name)
 
     @pytest.mark.tags(CaseLabel.L3)
