+++ tests/python_client/milvus_client/test_milvus_client_search_string.py
